Add more OOP
Grouped and abstracted logic and methods into new classes
