+++ src/main/data/duke.txt
@@ -1,1 +1,1 @@
-[D][ ] do hw (by: Oct 10 2022)
+[D][ ] do hw (by: Dec 12 2022)
