+++ text-ui-test/input.txt
@@ -3,4 +3,5 @@ list
 deadline sleep /by 2022-12-12
 event do tutorial /at 2022-08-30
 list
+hahahah
 bye
